% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlaspm-package.R
\docType{package}
\name{atlaspm-package}
\alias{atlaspm}
\alias{atlaspm-package}
\title{atlaspm: Bayesian Atlas of Preventable Mortality}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrea Pedot \email{andrea.pedot@unimi.it} (\href{https://orcid.org/0000-0001-7050-1378}{ORCID})

Authors:
\itemize{
  \item Andrea Pedot \email{andrea.pedot@unimi.it} (\href{https://orcid.org/0000-0001-7050-1378}{ORCID})
}

}
\keyword{internal}
