% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{collapse_codes}
\alias{collapse_codes}
\title{Collapse a set of ICD-10 keys into readable ranges}
\usage{
collapse_codes(keys)
}
\arguments{
\item{keys}{Character vector of ICD-10 prefixes.}
}
\value{
A length-1 string.
}
\description{
\code{c("A00","A01","A02","A05")} becomes \code{"A00-A02, A05"}.
}
\examples{
collapse_codes(c("A00", "A01", "A02", "A05"))
}
